CarRental/myproject/requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
djangorestframework==3.15.2
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
oauth2client==4.1.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyDrive==1.3.1
pyparsing==3.1.2
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.2
